# docs/api/swagger.yaml
definitions:
...
  domain.Paginator:
    description: Paginator represents a paginator.
    properties:
      limit:
        description: Maximum number of items per page
        example: 10
        format: int
        type: integer
      next_page:
        description: URL for the next page
        example: http://localhost:8080/users?next_token=ZmZmZmZmZmYtZmZmZi0tZmZmZmZmZmY=&limit=10
        format: string
        type: string
      next_token:
        description: Base64 encoded next page token
        example: ZmZmZmZmZmYtZmZmZi0tZmZmZmZmZmY=
        format: string
        type: string
      prev_page:
        description: URL for the previous page
        example: http://localhost:8080/users?prev_token=ZmZmZmZmZmYtZmZmZi0tZmZmZmZmZmY=&limit=10
        format: string
        type: string
      prev_token:
        description: Base64 encoded previous page token
        example: ZmZmZmZmZmYtZmZmZi0tZmZmZmZmZmY=
        format: string
        type: string
      size:
        description: Number of items in the current page
        example: 10
        format: int
        type: integer
    type: object
...
  payload.HTTPMessage:
    description: HTTPMessage represents a message to be sent to the client trough
      HTTP REST API.
    properties:
      code:
        description: |-
          Code is a stable, machine-readable reason, present only where a client
          has to BRANCH on it. Message is prose and may be reworded; this is not.

          It exists because a 401 has two meanings a client must treat differently:
          an expired access token should be refreshed and the request retried,
          while a revoked one must not be — the refresh token was revoked in the
          same breath, so the retry burns two requests and fails anyway. Nothing in
          the old response distinguished them, and a client cannot be asked to
          match on English.

          Empty on every response that does not need one, so no existing client
          sees a change.
        example: token_revoked
        format: string
        type: string
      message:
        example: success
        format: string
        type: string
      method:
        example: GET
        format: string
        type: string
      path:
        example: /api/v1/users
        format: string
        type: string
      status_code:
        example: 200
        format: int32
        type: integer
      timestamp:
        example: "2021-07-01T00:00:00Z"
        format: date-time
        type: string
    type: object
...
  payload.ListResourcesResponse:
    description: Paginated list of API resource permission definitions
    properties:
      items:
        description: Array of resource permission definitions
        items:
          $ref: '#/definitions/payload.ResourceResponse'
        type: array
      paginator:
        allOf:
        - $ref: '#/definitions/domain.Paginator'
        description: Pagination information including total count and page details
    type: object
...
  payload.ResourceResponse:
    description: API resource permission definition for authorization control
    properties:
      action:
        description: HTTP method or action type
        enum:
        - GET
        - POST
        - PUT
        - DELETE
        - PATCH
        example: GET
        format: string
        type: string
      created_at:
        description: Timestamp when the resource was created
        example: "2021-01-01T00:00:00Z"
        format: date-time
        type: string
      description:
        description: Detailed description of what this permission grants
        example: Allows reading of user data
        format: string
        type: string
      id:
        description: Unique identifier for the resource permission
        example: 019b4b0d-a682-7e48-b818-261829e39f76
        format: uuid
        type: string
      name:
        description: Human-readable name of the permission
        example: Read Users
        format: string
        type: string
      resource:
        description: API resource path or identifier
        example: /api/v1/users
        format: string
        type: string
      system:
        description: Indicates if this is a system-managed resource (cannot be deleted)
        example: false
        type: boolean
      updated_at:
        description: Timestamp when the resource was last updated
        example: "2021-01-01T00:00:00Z"
        format: date-time
        type: string
    type: object
...
info:
  contact:
    email: info@goapitemplate.local
    name: API Support
    url: https://goapitemplate.local
  description: |-
    A production-shaped starting point for a Go HTTP REST API: hexagonal architecture, multi-tenant projects, RBAC through Open Policy Agent, database-backed rate limiting and resource limits, and OpenTelemetry traces and metrics throughout.

    `products` is the worked example entity -- a project-scoped CRUD resource that exercises every convention here. Copy it when adding your own; see docs/architecture/adding-an-entity.md.

    ### Authentication

    Four bearer credentials, each accepted by a different set of endpoints; see the security definitions below for which is which. Ordinary calls use **AccessToken**.

    A logout revokes the access token it was called with, not only the refresh token, so it stops working immediately. That check can be switched off (`authn.access.token.revocation.enabled`), and where it is, a logged-out access token keeps working until it expires -- which is why the lifetime is short.

    Refresh tokens are rotated: each refresh spends the token it was given and returns a successor, so a client **must** store what `POST /auth/refresh` returns. Presenting a spent one again reads as a replay and ends the session for whoever holds the live token too.

    ### Conventions

    - **Identifiers are UUID v7.** Every `id` in this API, without exception.
...

    ### Rate limiting

...

...

...

    ### Operations

...
  version: v1
paths:
...
  /resources:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of all available system resources.
      operationId: 01982303-f0f9-7ee0-aa66-0f756c3c8bec
      parameters:
      - description: 'Sort order: comma-separated fields with ASC/DESC. Example: name
          ASC,created_at DESC'
        in: query
        name: sort
        type: string
      - description: 'Filter expression using SQL-like syntax. Example: action=''read''
          AND system=true'
        in: query
        name: filter
        type: string
      - description: 'Comma-separated list of fields to include in response. Example:
          id,name,action,resource'
        in: query
        name: fields
        type: string
      - description: Pagination cursor for fetching the next page of results
        in: query
        name: next_token
        type: string
      - description: Pagination cursor for fetching the previous page of results
        in: query
        name: prev_token
        type: string
      - description: 'Maximum number of items to return per page (default: varies
          by configuration)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated list of resources retrieved successfully
          schema:
            $ref: '#/definitions/payload.ListResourcesResponse'
        "400":
          description: Invalid query parameters or malformed filter/sort expression
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "401":
          description: Missing or invalid authentication token
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "429":
          description: Too many requests -- RATE_LIMIT_EXCEEDED is the budget, RATE_LIMIT_UNAVAILABLE
            the limiter's own store
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
      security:
      - AccessToken: []
      summary: List resources
      tags:
      - Resources
  /resources/{resource_id}:
    get:
      consumes:
      - application/json
      description: Retrieve detailed information about a specific system resource
        configuration using its unique identifier.
      operationId: 019822c9-9775-71b1-a2c6-deac83cf2519
      parameters:
      - description: Unique resource identifier (UUID v7)
        format: uuid
        in: path
        name: resource_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Resource details retrieved successfully
          schema:
            $ref: '#/definitions/payload.ResourceResponse'
        "400":
          description: Invalid resource ID format or malformed request
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "401":
          description: Missing or invalid authentication token
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "404":
          description: Resource not found
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "429":
          description: Too many requests -- RATE_LIMIT_EXCEEDED is the budget, RATE_LIMIT_UNAVAILABLE
            the limiter's own store
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
      security:
      - AccessToken: []
      summary: Get resource
      tags:
      - Resources
  /resources/matches:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of resources that match the specified
        action and resource policy patterns.
      operationId: 01982303-f0f9-7e44-aeb1-63934913e601
      parameters:
      - description: Action pattern to match (e.g., 'read', 'write', 'delete' or wildcard
          patterns)
        in: query
        name: action
        required: true
        type: string
      - description: Resource pattern to match (e.g., 'users/*', 'projects/123' or
          wildcard patterns)
        in: query
        name: resource
        required: true
        type: string
      - description: 'Sort order: comma-separated fields with ASC/DESC. Example: name
          ASC,created_at DESC'
        in: query
        name: sort
        type: string
      - description: 'Comma-separated list of fields to include in response. Example:
          id,name,action,resource'
        in: query
        name: fields
        type: string
      - description: Pagination cursor for fetching the next page of results
        in: query
        name: next_token
        type: string
      - description: Pagination cursor for fetching the previous page of results
        in: query
        name: prev_token
        type: string
      - description: 'Maximum number of items to return per page (default: varies
          by configuration)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Matching resources retrieved successfully
          schema:
            $ref: '#/definitions/payload.ListResourcesResponse'
        "400":
          description: Invalid action/resource pattern or malformed query parameters
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "401":
          description: Missing or invalid authentication token
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "404":
          description: No resources found matching the specified patterns
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "429":
          description: Too many requests -- RATE_LIMIT_EXCEEDED is the budget, RATE_LIMIT_UNAVAILABLE
            the limiter's own store
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
      security:
      - AccessToken: []
      summary: Find resources by action and pattern
      tags:
      - Resources
...
securityDefinitions:
  AccessToken:
    description: 'The ordinary credential: `Bearer <access_token>`, from `POST /auth/login`
      or `POST /auth/refresh`. Every endpoint outside `/auth`, `/health` and `/version`
      takes it. A personal access token is presented the same way and is accepted
      anywhere an access token is -- which is how a script authenticates without a
      login. Deliberately short-lived; refresh rather than lengthening it.'
    in: header
    name: Authorization
    type: apiKey
  RefreshToken:
    description: '`Bearer <refresh_token>`, and accepted **only** by `POST /auth/refresh`
      and `DELETE /auth/logout`. Refreshing spends it and returns a successor: store
      what comes back. Presenting a spent one is treated as a replay and ends the
      whole session, including for whoever holds the live token.'
    in: header
    name: Authorization
    type: apiKey
  ResetPasswordToken:
    description: '`Bearer <reset_password_token>`, accepted only by `POST /auth/password/reset`.
      Issued by `POST /auth/password/recover` and delivered by email, single use.'
    in: header
    name: Authorization
    type: apiKey
  VerificationToken:
    description: '`Bearer <verification_token>`, accepted only by `POST /auth/verify/confirm`.
      It proves an email address rather than exercising a permission, so it is the
      one credential no authorization check is applied to -- the account it belongs
      to is still disabled at that point. It never travels in a URL: the emailed link
      points at the frontend, which presents the token in this header.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...